Prevent running serialize/deserialize tests
xUnit tests in different classes run in parallel by default.  Try to read and write the XML file from different classes was causing an error.
